Add CB_LIBFUZZER
Add support for using LibFuzzer (https://llvm.org/docs/LibFuzzer.html)
in the build process by passing -DCB_LIBFUZZER=ON when running
CMake.

That'll add "-fsanitize=fuzzer-no-link" to the compile line for source
files, and for the binaries which wants libfuzzer to install a main()
they need to do:

    cb_link_libfuzzer_main(name-of-target)

And the correct link options will be added for the target.

You would typically use CB_LIBFUZZER together with one of the
sanitizers, but it isn't a requirement.

Change-Id: I090c91be2fbadf03fc4d3c6320d481aa64e4c105
Reviewed-on: http://review.couchbase.org/c/tlm/+/156676
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>