MB-46988: Fix Windows LNK4221 warning after adding PCH support
After adding precompiled header support, specifically
"MB-46988: Fix PCH headers causing ccache cache misses"(1da8a18),
we see the following linker error when building with MSVC:

    cmake_pch.cxx.obj : warning LNK4221: This object file does not
    define any previously undefined public symbols, so it will not
    be used by any link operation that consumes this library

This is due to the fact we end up with the PCH object file
(cmake_pch.cxx.obj) on the link line twice.

This is a result of the aforementioned patch adding the following line
to the reuse_pch() function:

    target_link_libraries(${target} PRIVATE ${pch_target})

Note that 'pch_target' is already linked to 'target' via
the target_precompile_headers() call, and for whatever reason CMake
doesn't de-duplicate them.

Fix this by solving the original ccache miss issue in a slightly
different way - instead of linking to the entire PCH library, just
lookup it's INCLUDE_DIRECTORIES (and those of it's dependancies)
and add those to the target's INCLUDE_DIRECTORIES.

Change-Id: Ibbcbeb757f532d56e35fcb042572076ab622efa8
Reviewed-on: http://review.couchbase.org/c/tlm/+/156873
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>