CBD-3564: Restore grpc as a v1 cbdep.
This was originally removed when converting grpc to v2 cbdeps and
creating a new v2 grpc 1.31.1 package for SDK. However, Server still
uses the v1 grpc 1.28.1 package, so this code should remain here until
it upgrades. (In particular we want to update it to build for aarch64.)

This partially reverts commit e343252f797c1eab97667bb834c90ac22e689ce4.

Change-Id: I584378571611ce872a9ab2a9e67483fd67d4f6ba
Reviewed-on: http://review.couchbase.org/c/tlm/+/157276
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>