MB-47864: Use rebuild jemalloc 5.2.1-cb6 for centos8
When 5.2.1-cb5 was built for MB-46126, centos8 was overlooked, and
unfortunately an older 5.2.1-cb5 was picked up. 5.2.1-cb6 was created
later for CC and later builds, but it only has changes relevant to MacOS
and aarch64. So in the interest of minimal change in 6.6.3, I have built
5.2.1-cb6 for centos8 and am changing only centos8 builds in 6.6.3 to
use that newer version.

Change-Id: I87a48931d8f501a0d7c105526bff19d5706bf2b7
Reviewed-on: http://review.couchbase.org/c/tlm/+/159108
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>