MB-43620: Snappy: disable BMI2 instructions
When compiling with MSVC, it will allow compiler intrinsics to be
emitted for instruction sets which are not actually enabled with the
current /arch setting. For example, the BMI2 instructions which are
part of the AVX2 instruction set can be emitted even if /arch:AVX2 is
not specified.

The consequence of this is that Snappy's checks to see if BMI2
instruction is available always succeed. There also doesn't appear to be
a flag to MSVC which will disable this functionality.

I couldn't find explicit documentation of this behaviour, but it is
somewhat implicitly documented at[1], which mentions the need to use
__cpuid to check at runtime if a particular (emitted) instruction is
actually available.

As such, explictly set SNAPPY_HAVE_BMI2 variable ahead of time, which
will skip the check and use the value we specify (NO).

[1]: https://docs.microsoft.com/en-us/cpp/intrinsics/x64-amd64-intrinsics-list?view=msvc-160

Change-Id: I6fa2b448384c3cbfab03784ab995d10323e3cdc3
Reviewed-on: http://review.couchbase.org/c/tlm/+/161098
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>