Allow to bypass third_party/googletest
To allow for testing with CV we need a way to bypass
inclusion of third_party/googletest. We can't do that
in the same patch as where we want to bypass it, as
it is repo which installs the CMakeLists.txt file and
that happens _before_ script to install the patch gets
executed.

By merging this patch we may upload additional patches
which set the value and we can bypass the directory

Change-Id: Icdce35c160779f035b4e1f4ec69dfb381dd090d5
Reviewed-on: http://review.couchbase.org/c/tlm/+/161243
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>