go-mod build: Only schedule one go-mod target at once (Ninja)
Go build/install already performs it's own parallelism internally, so
we don't also want to have the CMake generator attempt to parallelise
(i.e. run multiple `go build/install` targets in parallel). If we do
(particulary for machines which have large numbers of CPUs but perhaps
not as large RAM) then we can end up exhausing the RAM of the machine.

Define a CMake JOB_POOL which has concurrency 1, which is used by the
'go build' and 'go install' custom targets below.

This allows us to remove the "manual" serialisation of GoInstall
targets (by making each one depend on the previous one), which had the
downside of "forcing" CMake to always build all Go targets, even for
an incremental build.

Reduces no-op incremental build time from:

    $ hyperfine "ninja projector"
    Benchmark 1: ninja projector
      Time (mean ± σ):      3.354 s ±  1.232 s    [User: 2.711 s, System: 3.060 s]
      Range (min … max):    2.155 s …  6.025 s    10 runs

to:

    $ hyperfine "ninja projector"
    Benchmark 1: ninja projector
      Time (mean ± σ):     526.5 ms ±  16.5 ms    [User: 645.7 ms, System: 870.0 ms]
      Range (min … max):   508.3 ms … 566.5 ms    10 runs

As we actually only attempt to build the named target, and not all
other GoInstall targets also.

(Note: At time of writing this is only supported by the Ninja
generator, is it ignored by other generators.)

Change-Id: Ia8ac6307727635f4d6ad4de486a7aeb4e3d3c8e6
Reviewed-on: https://review.couchbase.org/c/tlm/+/161677
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>