MB-48220, CBD-4237: Improvements in cbpy packaging
We now create cbpy using "conda pack", which results in a plain
directory containing python that can be relocated. Because of this, we
lower the build and installer complexity since they don't need to call a
sub-installer for python.

This is installed in <install_dir>/lib/python/interp now, rather than
<install_dir>/lib/python/runtime, because previous version installers
make some assumptions about being able to delete the lib/python/runtime
directory. For rpm upgrades in particular, where the install/uninstall
steps are executed in a bizarre order when doing an upgrade, there was
no safe way to ensure upgrade would work from any version and always end
up with python still installed.

We also switched to Miniforge for our Python base, rather than
Miniconda. This is so we depend only on conda-forge and not Anaconda's
package repositories, which are now not available under favorable Terms
of Service. This change did result in some dependency versions being
slightly changed. We also now need to build python-snappy on Linux
aarch64 as that package is missing on conda-forge.

For the moment, the "local-cbpy" target is still here, because several
other parts of the Server build (correctly) depend on that target to
ensure PYTHON_EXE exists. That is no longer necessary, so those DEPENDS
can be removed; once they are, local-cbpy can be eliminated.

Change-Id: I088ce203685aab08d2ec1356ad6893c09e6798c5
Reviewed-on: http://review.couchbase.org/c/tlm/+/161737
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>