Sanitizers [1/3]: Pass ASan enablement flags via CMake properties
Update CouchbaseAddressSanitizer.cmake to match how
CouchbaseThreadSanitizer handles the flags needed to enable the
sanitizer; by passing the flags via properties on each target, instead
of globally modifying CMAKE_<LANG>_FLAGS.

The primary motiviation for this change (other than consistency with
TSan) is so we can subsequently disable Asan on a per-target basis for
Golang targets; which don't currently use the COMPILE_FLAGS /
LINK_FLAGS properties and hence would be difficult to disable TSan on
some Golang targets. This is done in the 3rd patch in the series.

Change-Id: I49bbab2bfefce6f151cfee41098ee03d1830ec2f
Reviewed-on: https://review.couchbase.org/c/tlm/+/163279
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>