MB-50252: Correctly pass ASan / UBSan flags as list to install_sanitizer_library
As part of updating ASan & UBSan CMake handling for Golang in:

  "Sanitizers [1/3]: Pass ASan enablement flags via CMake properties" (0db86b2a) and
  "Sanitizers [2/3]: Pass UBSan enablement flags via CMake properties" (403d71fe3c5)

The CMake variables ADDRESS_SANITIZER_FLAG and
UNDEFINED_SANITIZER_FLAG were changed to pass multiple flags as a
CMake list. Unfortunately this variable was not correctly quoted when
passed to install_sanitizer_library(), which resulted in the arguments
being parsed incorrectly, and the asan / ubsan runtime libraries
getting installed to the wrong locations:

    -- Found ASan at: /opt/gcc-10.2.0/lib64/libasan.so.6 installing to: -fno-omit-frame-pointer
    ...
    -- Found UBSan at: /opt/gcc-10.2.0/lib64/libubsan.so.1 installing to: -fno-sanitize=alignment

Fix by correctly quoting them.

Change-Id: Ifa7dfa9048e7b3a661b1c2189d39add7b72b050e
Reviewed-on: https://review.couchbase.org/c/tlm/+/168246
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>