CBD-4359: Update to Erlang 24
Erlang is now a cbdeps 2.0 package, and builds normally on Windows!!
(You do have to specify the platform "windows_msvc2019:wsl" to get
the right build agent.)

Also bump openssl to 1.1.1l-3 which includes static libraries on
Windows, which erlang requires.

Also change cbdeps v1 build processes to always create Windows packages
with only "windows" in the filename, rather than eg. "windows_msvc2017".
There is a parallel change for cbdeps v2 in build-tools. Having PLATFORM
include the compiler version is important for routing builds to the
right Jenkins agents, but not for the resulting package name.

Change-Id: I8fb576c1e798354c56dea795abdfc91fccdbaaeb
Reviewed-on: http://review.couchbase.org/c/tlm/+/163080
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>