MB-50252: Install GCC libs for non-GCC Linux compilers
+Background+

When using ThreadSanitizer, we build the C/C++ parts of CB Server with
Clang as it has generally given better thread analysis results
compared to GCC.

However, when using a Clang, we still want to use GCC's runtime
libraries for C++ programs - both libstdc++.so and libgcc_s.so. This
is essentially because the same C++ standard library needs to be used
across all object files within a process, and we have various
pre-compiled dependancies compiled with GCC and hence already linking
to libstdc++ (GCC's C++ standard library).

Given that we currently use a non-distro GCC installed into /opt/gcc
(version 10.2 as of time of writing), we consequently tell Clang to
use that GCC's standard libraries.  This is achieved by passing
'--gcc-toolchain=/opt/gcc-10.2.0' to Clang via CXXFLAGS env var.

+Problem+

While Clang can locate the /opt/gcc-10.2.0 runtime libraries at build
time, at runtime those libraries need to be in the runtime linker
search path for the compiled program to correclty execute. This is
done by installing the runtime libraries under our install prefix
(e.g. /opt/couchbase/lib) - see _install_gcc_file.

However, that macro is predicated on the compiler being GCC - which as
describd above is too restrictive; we also need to do it for Clang.

Fix by removing the check on the compiler being GCC.

We also hit a second issue - if we just invoke ${CMAKE_CXX_COMPILER}
then it does not include the --gcc-toolchain= flag - we also need to
include ${CMAKE_CXX_FLAGS} in the compiler command.

Change-Id: Ide22040b879fffac43520bf73bb060de39a14056
Reviewed-on: https://review.couchbase.org/c/tlm/+/168303
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>