MB-51532: Create jemalloc ReleaseAssertions build on Windows
This has jemalloc's debug assertions enabled.

Note that for non-Windows the debug variant is sufficient to be able
to run a normal Release/RelWithDebInfo with jemalloc runtime
assertions enabled; but on Windows we must link against the correct
CRT and hence need a third library variant
(Release CRT, but debug assertions).

Change-Id: I4d7710e0d38a0120f2a5e4fc7bd62f9b0980db3d
Reviewed-on: https://review.couchbase.org/c/tlm/+/172652
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>