MB-51532: Have CMake treat 'DebugOptimized' as a debug build type
'DebugOptimized' is a custom build type we added which is based on
Debug but with a small amount of optimisation enabled to speed up
runtime of CV jobs.

However, it was not recognised by CMake as a debug build type, so if
there were different versions of a library available for debug vs
optimized builds, we would use the optimized one for DebugOptimized.

A forthcoming patch adds a debug build of the jemalloc library which
we want DebugOptimized to use, so include DebugOptimized in the set of
build types CMake considers 'debug'.

Change-Id: I69c2c536e37bbd843599f1a667e371406a9b4b84
Reviewed-on: https://review.couchbase.org/c/tlm/+/172964
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>