MB-51719: Build fmtlib v8.1.1
Create a cbdeps of fmtlib v8.1.1, but don't yet use it. This is
required to upgrade Folly to latest stable (requires at least 8.0.1).

We are not yet using this version of fmtlib, as we also need to
rebuild spdlog to use the new version; if tried to use v8 of fmtlib
then the current version of spdlog fails to build. After spdlog has
also been rebuilt for the newer version, we can switch to use both
fmtlib v8.1.1 and spdlog v1.10.0 simultaneously.

Change-Id: I57f91d00e82dda88323322081fec78fc0f73c3cf
Reviewed-on: https://review.couchbase.org/c/tlm/+/173281
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>