MB-51722: Upgrade spdlog to v1.10.0 and switch to fmtlib v8.1.1-cb2
***
Re-apply after revert, due to broken fmt cbdep package being
uploaded. This is identical to 21cc26f apart from:
1. Bumping the version of fmt to cb2, and then consequently
   bumping spdlog's build version (as it now uses fmt-cb2).
2. Maintaining the platform specific packages instead of the
   new "linux" package (which has problems with TSan variant
   - see CBD-4186).
***

We need to upgrade spdlog to at least v1.9 to support fmtlib 8.x -
which is needed to upgrade Folly.

Note we must switch to spdlog v1.10 and fmtlib 8.1.1 simultaneously as
they spdlog 1.10 requires fmtlib v.8.x - it fails to compile with the
previous v7.

As such, this patch builds spdlog against the (already built) fmtlib,
and also switches the manifest to use both in the same commit.

Change-Id: I62168f7eeb8a3f125f1028e39114b344da8046d9
Reviewed-on: https://review.couchbase.org/c/tlm/+/173972
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>