MB-51722: Upgrade spdlog to v1.10.0 and switch to fmtlib v8.1.1
We need to upgrade spdlog to at least v1.9 to support fmtlib 8.x -
which is needed to upgrade Folly.

Note we must switch to spdlog v1.10 and fmtlib 8.1.1 simultaneously as
they spdlog 1.10 requires fmtlib v.8.x - it fails to compile with the
previous v7.

As such, this patch builds spdlog against the (already built) fmtlib,
and also switches the manifes to use both in the same commit.

Change-Id: I244986ec03ff5bcdf4aad600df62773ddd29d986
Reviewed-on: https://review.couchbase.org/c/tlm/+/173286
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>