CB_GET_SUPPORTED_PLATFORM: Consider 'linux' a supported platform
Given we now use the platform 'linux' to refer to the single-linux
cbdep, and such a platform _is_ supported, it should included in the
set of platforms that CB_GET_SUPPORTED_PLATFORM checks for.

Note this is required when building a v1 cbdep which is building on
'linux' and that cbdep itself depends on another cbdep which specifies
the 'linux' version of that dependancy.

Change-Id: If67436b27b616802debb56c5741ffe0b20dc6575
Reviewed-on: https://review.couchbase.org/c/tlm/+/173349
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>