MB-51718: Allow specific or generic linux packages during cbdeps build
When building a cbdep which itself requires some cbdep packages
(e.g. Folly), the current method of creating single-linux package
(platform="linux") requires that _all_ of the dependancies also are
for the specific "linux" distro. For example, I cannot do the
following inside deps/packages/folly/CMakeList.txt (if I am also
trying to build for the single-linux platform):

    DECLARE_DEP (boost VERSION 1.74.0-cb1 PLATFORMS amzn2 centos7 centos8 ...)

The problem is that cbdep fails to match this DECLARE_DEP against the
specified platform:

    cmake .. -DPACKAGE=folly '-DCB_DOWNLOAD_DEPS_PLATFORM=linux;all'
    <cut>
    -- Dependency glog (v0.4.0-cb1) not declared for platform linux;all;all, skipping...
    -- Dependency jemalloc (5.2.1-cb6) not declared for platform linux;all;all, skipping...

The problem is that we currently override CB_DOWNLOAD_DEPS_PLATFORM -
ie. the platform this machine identifies as - to be "linux". Given
that is not listed in the above DECLARE_DEP() macros, we don't find
the required dep.

The only current workaround for this would be to rebuild and de-define
_all_ DECLARE_DEPs inside deps/packages/folly/CMakeList.txt as

    DECLARE_DEP (boost VERSION x.y.x PLATFORMS linux)
    ...

However this feels broken - when the "normal" server build consumes
dependancies it can use a mix of generic (linux) and specific
(centos7) packages, but we cannot here.

The problem here is that we are changing platform in the wrong place -
we shouldn't be changing what platform we identify as (and hence what
platform(s) we can consume from), we should just be changing that
platform we are producing packages for.

As such, address this by removing the override of
CB_DOWNLOAD_DEPS_PLATFORM - so we identify as say "centos7" and hence
can use both "centos7" *and* "linux" packages; but then allow an
override of what the resulting output cbdeps file is named -
CB_DEP_TARGET_PLATFORM.

Change-Id: Ic1a4d95a756d2f4ee8dcefe1066075c903cf0fbc
Reviewed-on: https://review.couchbase.org/c/tlm/+/173488
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>