Revert "MB-51722: Upgrade spdlog to v1.10.0 and switch to fmtlib v8.1.1"
The binary package of fmt-v8.1.1-cb1 has an incorrect "lib64/" subdir
(should be "lib/") on the Linux build. This issue is fixed in the
CMakeLists.txt during fmt v8.1.1 upgrade, however, a previous build of
the cbdep had already been promoted (and subsequently published).

Revert back to previous spdlog / fmtlib versions so a -cb2 build of
fmtlib can be created.

This reverts commit 21cc26fc12570440de39c348dedd663596120ecf.

Change-Id: I681c79026a731df1ecf93fde467f9048024761de
Reviewed-on: https://review.couchbase.org/c/tlm/+/173971
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>