MB-52856: Revert "MB-51532: Have CMake treat 'DebugOptimized' as a debug build type"
As part of MB-51532, the debug variant of jemalloc (libjemallocd.so)
was used for debug build types - "Debug" and "DebugOptimized". This
has resulted in a significant increase in the runtime of CV tests -
for example on macOS (Apple M1 Pro), DebugOptimized build the total
runtime for all kv_engine tests is now:

    Total Test time (real) = 518.24 sec

    real     8m38.260s
    user    61m50.453s
    sys      3m2.256s

Compared to before when DebugOptimized used the non-debug variant of jemalloc:

    Total Test time (real) = 177.53 sec

    real     2m57.544s
    user    16m30.490s
    sys      2m44.285s

i.e user+sys has increased from 19m14s to 64m42s - over 3x slower.

While enabling extra debug checks for DebugOptimized can be useful,
having it cost 3x in test runtime is not really an acceptable
trade-off.

(Note jemalloc debug library is _still_ used for a straight "Debug"
build, so it can still be used if necessary).

This reverts commit ad8e4829e04ec655ae174270d6356aaa7b947719.

Change-Id: I33c216d1d6057e6feef54949dc93bf967db52ea9
Reviewed-on: https://review.couchbase.org/c/tlm/+/177146
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>