Add FATAL_ERROR with message when building for Apple Silicon
Add FATAL_ERROR with message when building for Apple Silicon as
currently we don't support it, and cmake won't fail with a nice error
message.

There's a nasty sublety with CMake here - if you attempt to run cmake
a second time (after seeing the error message about using the wrong
arch and dutifully setting the variables the message tells you to, the
build _still_ has CMAKE_SYSTEM_PROCESSOR set to arm64, because cmake
actually caches this under CMakeFiles/<VERSION>/CMakeSystem.cmake
which contains entries of the form:

    set(CMAKE_SYSTEM "Darwin-21.5.0")
    set(CMAKE_SYSTEM_NAME "Darwin")
    set(CMAKE_SYSTEM_VERSION "21.5.0")
    set(CMAKE_SYSTEM_PROCESSOR "arm64")

If this file is present then cmake ignores the value of
CMAKE_APPLE_SILICON_PROCESSOR passed on the command line. To avoid
this problem and allow a user to reconfigure following our
instructions, we must delete that file as part of printing the fatal
error message.

Change-Id: Ia7e884c3074c23356bf24bc5863c2f8108492bef
Reviewed-on: https://review.couchbase.org/c/tlm/+/177936
Tested-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>