CBD-5146: Improve go.mod developer experience
go-modbuild.cmake checks whether the main go.mod file is in a git
directory that has any local changes. If it doesn't, we presume it's an
"other" project that is not specifically being developed by the current
developer. Copy the go.mod file to a temp location, and build with 'go
build -mod=mod -modfile=/path/to/temp-go.mod'. This effectively makes
the go.mod file read/write but does not save the changes in a place git
can see them.

If there are local git changes, we now build with 'go build -mod=mod'
which allows the Go compiler to update go.mod/go.sum automatically,
without requiring the devs to run an explicit go mod tidy.

Change-Id: I2ffb2ec2e67855b56dcbb8c2b3aacaaefc7d4a76
Reviewed-on: https://review.couchbase.org/c/tlm/+/180425
Reviewed-by: Dhruvil Shah <dhruvil.ketanshah@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>