CBD-5142: Centralize management of 'fake' private Go modules
We hack in fake Go modules (just a directory with an empty go.mod) for
all private modules that are not accessible when syncing only source-
available projects in the manifest. Previously that was managed by
individual projects, but that's not ideal; everybody needs the same
potential set of faked modules, so it makes sense to keep a single
centralized list.

This change does that, and temporarily replaces the GoPrivateMod()
macro with an implementation that throws a CMake warning. We can
remove the macro entirely when the query and cbft projects (the only
current users of that macro) remove their calls to it.

Change-Id: I9fbfb26a18c99951a070cd696092d614cdd5cdbd
Reviewed-on: https://review.couchbase.org/c/tlm/+/180635
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>