Automatically set UBSan suppressions file when running tests
We already have a suppressions file for UBSan in
tlm/ubsan.suppressions, but it is not automatically enabled when
running unit tests - only during CV (see
https://review.couchbase.org/c/server-cv/+/173821). Update the CMake
scripts for sanitizers to automatically set the appropriate env vars
to enable the suppressions file for UBSan like we already do for ASan
& TSan.

To achieve this a restructure of how we setup tests for the various
sanitizers was needed:

Previously ASan & TSan each defined their own ADD_TEST() function
which overrode the default ADD_TEST() and set the appropriate env vars
for each sanitizer. This worked fine as ASan & TSan are mutually
exclusive - you cannot build with both enabled. his is not the case
with ASan & UBSan - indeed we commonly build with both at once as we
can build once and get checks from both sanitizers.

As such, move the overridden ADD_TEST() function to the general
CouchbaseSanitizers.cmake file, creating sanitizer-specific functions
to set the appropriate env vars.

Change-Id: If5a179589f1800fb82be8d8c09de302e37b29a8b
Reviewed-on: https://review.couchbase.org/c/tlm/+/180954
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>