MB-54162: cbdeps: build bebug variant of Folly for non-Windows platforms
We currently build a Debug variant of Folly for Windows, as it is
required there for the Debug build of CB Server (due to there being
different versions of the CRT for Debug / Release builds).

However, it is also useful to build for other platforms (Linux &
macOS); as the Debug build enables many internal assertions which can
ensure we are using Folly correctly.

Change-Id: Ic49c1d2f49f209730ada08ee1e62faf80fbd533a
Reviewed-on: https://review.couchbase.org/c/tlm/+/181413
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>