MB-54162: cbdeps: Build folly with same compiler as Server
As part of building a Debug variant of Folly, we need to build Folly
with the same compiler as the Server.

Specifically, we need to use a compiler which has support for C++17
std::filesystem, which GCC-10.2 does and GCC 7.3 does not, otherwise
we run into linker issues where libfolly.a (with GCC 7.3) has been
compiled to use std::experimental::filesystem (pre-C++ 17 namespace)
and CB Server has been compiled to use std::filesystem.

Change-Id: I4c53c9024883216aefe636bd7f8b86f2abde17eb
Reviewed-on: https://review.couchbase.org/c/tlm/+/181471
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>