MB-54275: Fix detection of updated cbdeps VERSIONs containing '-'
The function DECLARE_DEP assumes that only one hyphen will be present
in the version string when extracting the build number. However this
is not the case with folly and prometheus-cpp - for example:

    DECLARE_DEP (folly VERSION v2022.05.23.00-couchbase-cb6 PLATFORMS ...
    DECLARE_DEP (prometheus-cpp VERSION v0.10.0-couchbase-cb2 PLATFORMS ...

As such, the actual build number ends up getting ignored when
calculating the version - which is persisted as VERSION.txt:

    $ cat build/tlm/deps/folly.exploded/VERSION.txt
    v2022.05.23.00-couchbase

As such, when checking if a cbdep has already been extracted, the
check only compares the upstream version and not the cbdep build
number; resulting in new versions of a dependancy not getting
extracted - this is what is printed if folly is changed from build cb4
to cb6 - i.e. nothing is extracted:

    -- Checking exploded folly version v2022.05.23.00-couchbase against v2022.05.23.00-couchbase
    -- Dependency 'folly (v2022.05.23.00-couchbase)' already downloaded

Fix by removing the assumption about how many hyphens are present, and
simply splitting the string on the last hypen in it. If no hyphen then
the build number is empty. Build number is now parsed correctly, and
new version is extracted:

    -- Checking exploded folly version v2022.05.23.00-couchbase against v2022.05.23.00-couchbase-cb6
    -- Installing dependency: folly-v2022.05.23.00-couchbase-cb6...
    -- Extracting /home/couchbase/.cbdepscache/folly-ubuntu18.04-x86_64-v2022.05.23.00-couchbase-cb6.tgz to /home/couchbase/server/master/build/tlm/deps/folly.exploded.temp
    -- Moving /home/couchbase/server/master/build/tlm/deps/folly.exploded.temp to /home/couchbase/server/master/build/tlm/deps/folly.exploded

Change-Id: I7b065634525de62eec8844f16bbb11806119eb84
Reviewed-on: https://review.couchbase.org/c/tlm/+/181881
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>