Set CMAKE_OSX_DEPLOYMENT_TARGET to 12 for arm64
All the cbdeps for macOS/arm64 have been built on Monterey, and hence
if one tries to link to them with a lower CMAKE_OSX_DEPLOYMENT_TARGET
then warnings are such as this are emitted:

    [216/229] Linking CXX executable kv_engine/kvstore_gen
    ld: warning: dylib (tlm/deps/zlib.exploded/lib/libz.dylib) was built for newer macOS version (12.0) than being linked (11.0)
    ld: warning: dylib (../install/lib/libjemalloc.dylib) was built for newer macOS version (12.0) than being linked (11.0)
    ld: warning: dylib (tlm/deps/openssl.exploded/lib/libssl.dylib) was built for newer macOS version (12.0) than being linked (11.0)
    ...

Change CMAKE_OSX_DEPLOYMENT_TARGET for arm64 to 12.0 to avoid this.

(Recommit after previous revert due to relying on an incorrect CMake
variable (CMAKE_APPLE_SILICON_PROCESSOR) which is not set on x86_64
and not set before project() command).

Change-Id: Ia599be43bd8b439c70c370580b625d59f41fd3e2
Reviewed-on: https://review.couchbase.org/c/tlm/+/182320
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>