MB-57962: Folly cbdep: Create ASan+UBSan build variant
When building with ASan/UBSan on AppleClang 13.1.6 we see spurious
ASan errors reported when running binaries
(e.g. ep-engine_ep_unit_tests) linked against the vanilla build of
Folly. This appears to be due to the mismatch in some objects having
sanitizer annotations and some not in the same binary. Rebuilding
Folly with ASan+UBSan enabled addresses this issue.

As such, add an additional 'asan_ubsan' variant of Folly library which
is built for macOS & Ubuntu (CV platform), and update CMake scripts to
pickup this variant when ASan/UBSan are enabled.

Change-Id: I21cf6d2198850c1aec80386f463753a1768003ea
Reviewed-on: https://review.couchbase.org/c/tlm/+/183437
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>