MB-55268: [BP] Use forked jemalloc with --with-lg-tcache-limit=15
Backport change to 7.1.4-MP2.

This will increase the maximum number of tcaches to 32K and reduce the
maximum number of arenas to 512. This should be sufficient for our
current official limit of 30 buckets.

We use a tcache per bucket, per thread. Without this change, we'd hit
the limit of 4K tcaches easily on a 64-core machine with 20 Magma
buckets.

Change-Id: Ia16eb29b4d2354cbaec2a008a169c5de9c7a2afa
Reviewed-on: https://review.couchbase.org/c/tlm/+/189502
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>