MB-55268: Use forked jemalloc with --with-lg-tcache-limit=15
This will increase the maximum number of tcaches to 32K and reduce the
maximum number of arenas to 512. This should be sufficient for our
current official limit of 30 buckets.

We use a tcache per bucket, per thread. Without this change, we'd hit
the limit of 4K tcaches easily on a 64-core machine with 20 Magma
buckets.

Change-Id: I2d010e04b52ca56528307da64b7ccefb2ae01dab
Reviewed-on: https://review.couchbase.org/c/tlm/+/187077
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>