Add convenience script to build the server
Add a convenience script to build the server by using Ninja as
the build tool. The goal is to get rid of the top level Makefile
due to the problem with restricting certain job types (in addition
Ninja automatically determines the parallelism)

If invked without any arguments it behaves exactly like executing
make without any arguments in the root of the source directory.
It creates a sub-directory named build where the object files
resides (and you may build additional components), and the built
server gets installed into a sub-directory named install.
This means that you could just set the PATH variable to include
the location of the script and use it to build old versions
of Couchbase Server which don't contain the script.

You may tweak the build by passing command line options to
the script (or set environment variables):

Usage:

    -s source_root  Source directory (default: ${source_root})
    -b build_root   Build directory (default: ${build_root})
    -i install_root Install directory (default: ${install_root})
    -X              Set Mac platform to x86_64 (Only needed when
                    building on Mac running arm64)
    -T              Enable thread sanitizer
    -A              Enable address sanitizer
    -U              Enable undefined behavior sanitizer
    -R              Set build type to RelWithDebInfo

The following environment variables may be used to tweak
other settings:

* `CB_PARALLEL_LINK_JOBS` The number of link jobs to run in parallel.
  Due to static linking (and C++) each link process may consume *a lot*
  of memory (I've seen it go beyond 1GB). The default is set to two
  link steps in parallel.
* `EXTRA_CMAKE_OPTIONS` Extra options you would like to pass on to
  cmake.
* `CMAKE_BUILD_TYPE` By default it produce a `Debug` build; this may be
  set to `Debug`, `Release`, `RelWithDebInfo` or `MinSizeRel`

Finally, additional arguments may be passed to ninja like:

    trond@ok ~> ./Build.sh -- --option-to-ninja

Change-Id: I38d4d8eac5dc04ecdfbda0c5ec0af07d1c7a247d
Reviewed-on: https://review.couchbase.org/c/tlm/+/192994
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>