MB-57824: gtest_discover_tests: Fix passing of multiple env vars
(Backport to fix Asan issues on neo branch under MB-59356)

+Background+

For tests registered via gtest_discover_test(), we cannot rely on
add_sanitizers() to set the necessary sanitizer env vars (via
add_sanitizer_env_vars_memory et al), because the set of test names is
generated at runtime and hence CMake cannot know the name of the test
target to set properties on.

Instead we must pass the environment variables which need to be set to
gtest_discover_tests() via the ENVIRONMENT property.

+Problem+

When passing multiple env vars via the ENVIRONMENT property, they are
not correctly propogated all the say to GTest, and we end up losing
all but the first env var.

This is due to CMake's lack of proper list typing, the ENVIRONMENT
"sub-list" in the properties gets expanded and treated as additional
properties - e.g. the PROPERTIES:

   TIMEOUT 600 ENVIRONMENT TSAN_OPTIONS=xxx ASAN_OPTIONS=yyy

are incorrectly expanded to three properties:

   TIMEOUT:600
   ENVIRONMENT:TSAN_OPTIONS=xxx
   ASAN_OPTIONS=yyy

instead of what we need:

   TIMEOUT:600
   ENVIRONMENT:TSAN_OPTIONS=xxx;ASAN_OPTIONS=yyy

+solution+

This is a bit of a challenge to fix; we need to avoid having the
ENVIRONMENT property list be expanded until it is actually passed to
CTest - i.e. we need to stop the list elements from getting split.

CMake does have a facility to help with this problem in 3.0 - "backet"
arguments which act like pyhton or C++ raw strings and do not expand
anything betwen the delimiters [==[ and ]==] until the argument is
finally evaluated. However this doesn't work by itself here as
gtest_discover_test() *itself* wraps parameters in bracket arguments
if they contain any "unsafe" characters - which include the '[' and
'=' used for the bracket arg prefix / suffix...

As such, it is necessary to also modify gtest_discover_test's helper
functions and not re-escape arguments which already are.

 With both these changes we can successfully pass multple env vars
through:

    264: Test command: /cb/build-asan-ubsan-arm64/kv_engine/ep-engine_ep_unit_tests "--gtest_filter=RocksFullOrValue/DurabilityWarmupTest.*"
    264: Working Directory: /cb/build-asan-ubsan-arm64/kv_engine
    264: Environment variables:
    264:  TSAN_OPTIONS=suppressions=/cb/tlm/tsan.suppressions second_deadlock_stack=1 history_size=7 detect_deadlocks=1
    264:  LSAN_OPTIONS=suppressions=/cb/tlm/lsan.suppressions
    264:  ASAN_SYMBOLIZER_PATH=/usr/bin/llvm-symbolizer

Change-Id: I1db6482974584242fb3923d0bc1ca701e982f881
Reviewed-on: https://review.couchbase.org/c/tlm/+/200472
Well-Formed: Restriction Checker
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>